% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_summary_dataset.R
\docType{data}
\name{earthquake_summary_dataset}
\alias{earthquake_summary_dataset}
\title{Data summarising the earthquakes affecting each region, grouped by year}
\format{
A tibble with 37 rows and 5 variables:
\describe{
\item{City}{chr city of the recorded earthquakes}
\item{Year}{int year of the recorded earthquakes}
\item{Number of Earthquakes}{int total number of earthquakes greater than magnitude 4.0 that occured}
\item{Average Magnitude}{dbl average magnitude of earthquakes greater than magnitude 4.0}
\item{Max Magnitude}{dbl highest recorded magnitude in the given year and region}
}
}
\source{
\url{https://quakesearch.geonet.org.nz/}
}
\usage{
earthquake_summary_dataset
}
\description{
Includes every earthquake of a magnitude greater than 4.0
}
\keyword{datasets}
