% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mental_health_dataset.R
\docType{data}
\name{mental_health_dataset}
\alias{mental_health_dataset}
\title{Data of Uptake of Mental Health Services in New Zealand}
\format{
An object of class \code{data.frame} with 129 rows and 5 columns.
}
\source{
\url{https://www.health.govt.nz/nz-health-statistics/health-statistics-and-data-sets/mental-health-data-and-stats}
}
\usage{
mental_health_dataset
}
\description{
Shows the number of people accessing mental health services by major region in New Zealand
}
\details{
@format A tibble with 129 rows and 4 variables:
\describe{
\item{City}{chr city of recorded value, either Auckland, Canterbury or Wellington}
\item{Year}{int year of total, ranging from 2005 to 2020}
\item{Sex}{chr sex of the recorded value, either Female, Male or Total}
\item{Total}{int total number of people accessing Mental Health Services}
}
}
\keyword{datasets}
