% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzmentalquake-package.R
\docType{package}
\name{nzmentalquake-package}
\alias{nzmentalquake}
\alias{nzmentalquake-package}
\title{nzmentalquake: Title}
\description{
This package contains data on earthquakes, mental health service access and population of the three largest NZ regions - Auckland, Wellington and Canterbury.
}
\author{
\strong{Maintainer}: Chenthi Heer \email{chenthiheer@gmail.com} (\href{https://orcid.org/0000-0002-7694-0723}{ORCID})

}
\keyword{internal}
