% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dataset.R
\docType{data}
\name{population_dataset}
\alias{population_dataset}
\title{Data of Total Annual Population by Major Region in New Zealand}
\format{
An object of class \code{data.frame} with 48 rows and 5 columns.
}
\source{
\url{http://infoshare.stats.govt.nz/Default.aspx#}
}
\usage{
population_dataset
}
\description{
Simple population data
}
\details{
@format A tibble with 48 rows and 4 variables:
\describe{
\item{year}{int year of the recorded total, ranging from 2005 to 2020}
\item{City}{chr city of recorded population, either Auckland, Canterbury or Wellington}
\item{Total.Pop}{int the total population of the city in the given year}
\item{Pop.change}{int the difference in population from the previous year in the given city}
}
}
\keyword{datasets}
